Move runtimeconfig resources to the service package (#7933)
* Move runtimeconfig_config

* Move resource google_runtimeconfig_variable

* Move data sources

* Replace resources in provider

* Test PR 7930

* Import runtimeconfig package for beta

* Test gcb-terraform-vcr-tester:zhenhua-test-no-change

* Test gcb-terraform-vcr-tester:zhenhua-test-cloudtop

* Remove the testing image
