+++ google-beta/services/runtimeconfig/data_source_runtimeconfig_config.go
@@ -1,4 +1,4 @@
-package google
+package runtimeconfig
 
 import (
 	"fmt"
@@ -9,26 +9,25 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-func DataSourceGoogleRuntimeconfigVariable() *schema.Resource {
+func DataSourceGoogleRuntimeconfigConfig() *schema.Resource {
 
-	dsSchema := tpgresource.DatasourceSchemaFromResourceSchema(ResourceRuntimeconfigVariable().Schema)
+	dsSchema := tpgresource.DatasourceSchemaFromResourceSchema(ResourceRuntimeconfigConfig().Schema)
 	tpgresource.AddRequiredFieldsToSchema(dsSchema, "name")
-	tpgresource.AddRequiredFieldsToSchema(dsSchema, "parent")
 	tpgresource.AddOptionalFieldsToSchema(dsSchema, "project")
 
 	return &schema.Resource{
-		Read:   dataSourceGoogleRuntimeconfigVariableRead,
+		Read:   dataSourceGoogleRuntimeconfigConfigRead,
 		Schema: dsSchema,
 	}
 }
 
-func dataSourceGoogleRuntimeconfigVariableRead(d *schema.ResourceData, meta interface{}) error {
+func dataSourceGoogleRuntimeconfigConfigRead(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*transport_tpg.Config)
 
-	id, err := tpgresource.ReplaceVars(d, config, "projects/{{project}}/configs/{{parent}}/variables/{{name}}")
+	id, err := tpgresource.ReplaceVars(d, config, "projects/{{project}}/configs/{{name}}")
 	if err != nil {
 		return fmt.Errorf("Error constructing id: %s", err)
 	}
 	d.SetId(id)
-	return resourceRuntimeconfigVariableRead(d, meta)
+	return resourceRuntimeconfigConfigRead(d, meta)
 }
