+++ google-beta/services/runtimeconfig/resource_runtimeconfig_variable.go
@@ -1,4 +1,4 @@
-package google
+package runtimeconfig
 
 import (
 	"fmt"
